requirements.txt
anthropic>=0.40.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
sqlparse>=0.5.0
python-dotenv>=1.0.0
rich>=13.0.0